BackgroundRemoverOfObject/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
openai>=1.0.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
coloredlogs==15.0.1
distro==1.9.0
flatbuffers==25.12.19
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.14
pyparsing==3.3.1
python-dotenv==1.2.1
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-image==0.26.0
scipy==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.12.20
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
fastapi
uvicorn
python-dotenv
rembg
pillow
requests
replicate
fal-client
google-genai
uvicorn[standard]


